backend/requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Image processing
pillow>=10.0.0
numpy>=1.24.0
scikit-image>=0.21.0
opencv-python-headless>=4.8.0

# Model-specific libraries
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
hypothesis>=6.92.0
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
